Add periodic check to suppress packages from search results
After a discussion, we decided that many packages of the gddo database could be
suppressed from the search results. We are currently adopting 2 rules to
suppress a package:

1. Package project wasn't modified in the last 2 years and there're no other
projects with references to it.

2. Package project is a fork with a small number of commits near the fork date
(what we called a fast fork).

The periodic will check all the packages from the repository and send queries
to Github API to determinate the current state.

See #320